refactor: add refactor to tests
