--- test/unit/data/mq/mocks/reprocessing/index.ts
@@ -1,1 +1,0 @@
-export * from './save-reprocessing-data-repository-stub';
