+++ test/unit/data/protocols/db/reprocessing/index.ts
@@ -1,3 +1,4 @@
 export * from './get-reprocessing-data-by-identifier-repository-stub';
 export * from './delete-reprocessing-by-identifier-repository-stub';
 export * from './get-reprocessing-data-repository-stub';
+export * from './save-reprocessing-data-repository-stub';
