+++ test/unit/domain/index.ts
@@ -1,2 +1,2 @@
-export * from './examples';
-export * from './reprocessing';
+export * from './models';
+export * from './usecases';
