+++ test/unit/domain/models/examples/index.ts
