+++ test/unit/domain/models/index.ts
@@ -1,2 +1,2 @@
-export * from './mq';
+export * from './examples';
 export * from './reprocessing';
